e2e, tests: move the pod interface filter funcs to a dedicated pkg
