+++ e2e/retrievers/pod.go
@@ -1,0 +1,46 @@
+package retrievers
+
+import (
+	"encoding/json"
+	"fmt"
+
+	core "k8s.io/api/core/v1"
+
+	nettypes "github.com/k8snetworkplumbingwg/network-attachment-definition-client/pkg/apis/k8s.cni.cncf.io/v1"
+)
+
+func filterNetworkStatus(
+	networkStatuses []nettypes.NetworkStatus, predicate func(nettypes.NetworkStatus) bool) *nettypes.NetworkStatus {
+	for i, networkStatus := range networkStatuses {
+		if predicate(networkStatus) {
+			return &networkStatuses[i]
+		}
+	}
+	return nil
+}
+
+func SecondaryIfaceIPValue(pod *core.Pod) (string, error) {
+	podNetStatus, found := pod.Annotations[nettypes.NetworkStatusAnnot]
+	if !found {
+		return "", fmt.Errorf("the pod must feature the `networks-status` annotation")
+	}
+
+	var netStatus []nettypes.NetworkStatus
+	if err := json.Unmarshal([]byte(podNetStatus), &netStatus); err != nil {
+		return "", err
+	}
+
+	secondaryInterfaceNetworkStatus := filterNetworkStatus(netStatus, func(status nettypes.NetworkStatus) bool {
+		return status.Interface == "net1"
+	})
+
+	if secondaryInterfaceNetworkStatus == nil {
+		return "", fmt.Errorf("the pod does not have the requested secondary interface")
+	}
+
+	if len(secondaryInterfaceNetworkStatus.IPs) == 0 {
+		return "", fmt.Errorf("the pod does not have IPs for its secondary interfaces")
+	}
+
+	return secondaryInterfaceNetworkStatus.IPs[0], nil
+}
